+++ src/Context.php
@@ -527,8 +527,13 @@ abstract class Context
                 return false;
             }
 
-            // Short context name (must be formatted into class name).
-            $context = self::$contextPrefix . $context;
+            // Could be the fully qualified class name was given, like `ContextDBMS::class`.
+            if (class_exists('\\' . $context)) {
+                $context = '\\' . $context;
+            } else {
+                // Short context name (must be formatted into class name).
+                $context = self::$contextPrefix . $context;
+            }
         }
 
         self::$loadedContext = $context;